requirements/production.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --allow-unsafe --output-file=requirements/production.txt requirements/production.in
#
anyascii==0.3.1
    # via wagtail
argon2-cffi==21.3.0
    # via -r requirements/production.in
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.5.2
    # via django
beautifulsoup4==4.11.1
    # via wagtail
boto3==1.26.31
    # via -r requirements/production.in
botocore==1.29.31
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via
    #   argon2-cffi-bindings
    #   cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via typer
contourpy==1.0.6
    # via matplotlib
coreapi==2.3.3
    # via django-cast
coreschema==0.0.4
    # via coreapi
cryptography==38.0.4
    # via pyjwt
cyberpandas==1.1.1
    # via -r requirements/production.in
cycler==0.11.0
    # via matplotlib
defusedxml==0.7.1
    # via python3-openid
django==4.1.4
    # via
    #   django-allauth
    #   django-anymail
    #   django-braces
    #   django-cast
    #   django-contrib-comments
    #   django-extensions
    #   django-filter
    #   django-indieweb
    #   django-model-utils
    #   django-modelcluster
    #   django-permissionedforms
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   wagtail
django-allauth==0.51.0
    # via django-cast
django-anymail[mailgun]==8.6
    # via -r requirements/production.in
django-braces==1.15.0
    # via django-indieweb
django-cast @ git+https://github.com/ephes/django-cast@develop
    # via -r requirements/production.in
django-contrib-comments==2.2.0
    # via
    #   django-cast
    #   django-fluent-comments
    #   django-threadedcomments
django-crispy-forms==1.14.0
    # via
    #   django-cast
    #   django-fluent-comments
django-environ==0.9.0
    # via django-cast
django-extensions==3.2.1
    # via -r requirements/production.in
django-filter==22.1
    # via
    #   django-cast
    #   wagtail
django-fluent-comments==3.0
    # via django-cast
django-indieweb==0.0.6
    # via -r requirements/production.in
django-model-utils==4.3.1
    # via
    #   django-cast
    #   django-indieweb
django-modelcluster==6.0
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-proxy==1.2.1
    # via -r requirements/production.in
django-storages==1.13.1
    # via -r requirements/production.in
django-tag-parser==3.2
    # via
    #   django-cast
    #   django-fluent-comments
django-taggit==3.1.0
    # via wagtail
django-threadedcomments==2.0
    # via django-cast
django-treebeard==4.5.1
    # via wagtail
django-watson==1.6.2
    # via
    #   -r requirements/production.in
    #   django-cast
djangorestframework==3.14.0
    # via wagtail
draftjs-exporter==2.1.7
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
feedparser==6.0.10
    # via django-cast
fonttools==4.38.0
    # via matplotlib
gunicorn==20.1.0
    # via -r requirements/production.in
html5lib==1.1
    # via wagtail
idna==3.4
    # via requests
itypes==1.2.0
    # via coreapi
jinja2==3.1.2
    # via coreschema
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
kiwisolver==1.4.4
    # via matplotlib
l18n==2021.3
    # via wagtail
markdown==3.4.1
    # via django-cast
markupsafe==2.1.1
    # via jinja2
matplotlib==3.6.2
    # via -r requirements/production.in
numpy==1.23.5
    # via
    #   contourpy
    #   matplotlib
    #   pandas
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.0.10
    # via wagtail
packaging==22.0
    # via matplotlib
pandas==1.5.2
    # via
    #   -r requirements/production.in
    #   cyberpandas
pillow==9.3.0
    # via
    #   matplotlib
    #   wagtail
plotly==5.11.0
    # via django-cast
psycopg2==2.9.5
    # via -r requirements/production.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
pyparsing==3.0.9
    # via matplotlib
python-akismet==0.4.3
    # via
    #   django-cast
    #   django-fluent-comments
python-dateutil==2.8.2
    # via
    #   botocore
    #   matplotlib
    #   pandas
python-dotenv==0.21.0
    # via -r requirements/production.in
python-slugify==7.0.0
    # via django-cast
python3-openid==3.2.0
    # via django-allauth
pytz==2022.6
    # via
    #   django-indieweb
    #   django-modelcluster
    #   djangorestframework
    #   l18n
    #   pandas
requests==2.28.1
    # via
    #   -r requirements/production.in
    #   coreapi
    #   django-allauth
    #   django-anymail
    #   django-proxy
    #   python-akismet
    #   requests-oauthlib
    #   wagtail
requests-oauthlib==1.3.1
    # via django-allauth
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.12.0
    # via -r requirements/production.in
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   cyberpandas
    #   html5lib
    #   l18n
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.3
    # via django
telepath==0.3
    # via wagtail
tenacity==8.1.0
    # via plotly
text-unidecode==1.3
    # via python-slugify
typer==0.7.0
    # via -r requirements/production.in
uritemplate==4.1.1
    # via coreapi
urllib3==1.26.13
    # via
    #   botocore
    #   requests
    #   sentry-sdk
wagtail==4.1.1
    # via
    #   django-cast
    #   wagtail-srcset
wagtail-srcset==0.2.0
    # via django-cast
webencodings==0.5.1
    # via html5lib
whitenoise==6.2.0
    # via -r requirements/production.in
willow==1.4.1
    # via wagtail

# The following packages are considered to be unsafe in a requirements file:
setuptools==65.6.3
    # via gunicorn
